+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/view/customnavview/MapboxNavigationViewCustomizedActivity.kt
@@ -224,11 +224,7 @@ class MapboxNavigationViewCustomizedActivity : DrawerActivity() {
             menuBinding.toggleReplay,
             getValue = binding.navigationView.api::isReplayEnabled,
             setValue = { isChecked ->
-                if (isChecked) {
-                    binding.navigationView.api.enableReplaySession()
-                } else {
-                    binding.navigationView.api.enableTripSession()
-                }
+                binding.navigationView.api.routeReplayEnabled(isChecked)
             }
         )
     }